Have bigger histograms, remove other metrics entirely
The `uint16_t` -> `size_t` in `packedSize` is because now
insert stats requests are bigger than `uint16_t`.

+++ cpp/cdc/CDCDB.cpp
@@ -1187,7 +1187,7 @@ struct PackCDCReq {
         req.pack(buf);
     }
 
-    uint16_t packedSize() const {
+    size_t packedSize() const {
         return 1 + req.packedSize();
     }
 };
